# src/libkeccak: Use XKCP's non-CPU-specific 64-bit optimized implementation