--- a/.gitmodules
+++ b/.gitmodules
@@ -1,0 +1,3 @@
+[submodule "util/XKCP"]
+	path = util/XKCP
+	url = https://github.com/XKCP/XKCP.git