--- a/src/libkeccak-64/KeccakP-1600-opt64-config.h
+++ b/src/libkeccak-64/KeccakP-1600-opt64-config.h
@@ -1,0 +1,6 @@
+/*
+This file defines some parameters of the implementation in the parent directory.
+*/
+
+#define KeccakP1600_implementation_config "all rounds unrolled"
+#define KeccakP1600_fullUnrolling