… package.json
… src/README.md
… src/libkeccak-32/KeccakP-1600-inplace32BI.c
… src/libkeccak-64/KeccakP-1600-64.macros
… src/libkeccak-64/KeccakP-1600-opt64.c
… src/libkeccak-64/KeccakP-1600-unrolling.macros
… src/libkeccak-64/KeccakSponge.inc
… src/libkeccak-64/SnP-Relaned.h
… src/libkeccak-64/brg_endian.h
… src/libkeccak/KeccakP-1600-reference.c
… src/libkeccak/KeccakP-1600-reference.h
… src/libkeccak/KeccakSponge.c
… src/libkeccak/KeccakSponge.h
… test/vectors-sha3.js
… test/vectors-shake.js
… util/libkeccak.sh