It looks like we were using the reference C implementation, which is not
designed for good performance.

Switching to the non-CPU-specific 64-bit optimized implementation makes
us ~2x faster on small inputs and ~6x faster on large inputs.