@fanatid 2018-12-19
@cakoose users with 32-bit CPU will receive segfault?