@cakoose 2018-12-19
I think it will run correctly, but it will not run as quickly as the 32-bit optimized code.

Unfortunately, I'm on macOS and can't test because I can't find a 32-bit build of Node.

If you're on Linux, I think you can install a 32-bit build of Node ([link](https://nodejs.org/download/release/v10.14.2/)) and then do:
```
$ node-gyp clean configure build --verbose --arch=ia32  # force 32-bit build
...
$ file build/Release/keccak.node  # confirm that binary is 32-bit
...
$ npm test
...
```

Would also be interesting to run the benchmark on 32-bit node and see if the current reference code runs faster than the 64-bit optimized code.

Ideally we would provide both the 32-bit and 64-bit optimized code and have node-gyp select the code depending on the platform.  It seems like that should be possible, but unfortunately I know very little about node-gyp.