@fanatid 2018-12-19
After https://github.com/cakoose/keccak/pull/1 we will need figure out how support 32-bit version and can release major version.