@cakoose 2018-12-19
I figured out how to configure "binding.gyp" to do what we want.
1. I imported both the 32-bit optimized C code and the 64-bit optimized C code from XKCP.
2. If the Node `process.arch` is "arm64", "ppc64", or "x64", we will build the 64-bit optimized implementation.
3. Otherwise, we will build the 32-bit optimized implementation.

NOTE: I tested the 32-bit optimized C code on my 64-bit machine and it still runs faster than the "reference" C code we currently use (1.7x faster on short inputs, 2.5x faster on long inputs).