@fanatid 2018-12-19
Awesome! Thank you @cakoose 
Please check https://github.com/cakoose/keccak/pull/1