@cakoose 2018-12-19
I merged your commits into my branch and the tests passed.  (I also incorporated your "src/README.md" fix into my own commit.)

Is there anything else you would like me to do?