@fanatid 2018-12-20
Thank you for big update and performance improvement! Published as 2.0.0